udpnet: ignore source UDP port in multicast rx
On reception of a UDP datagram, check if it is coming from a multicast
service. That is, check if the source IP address matches with the IP address
defined in the udpmulticast conf instance and if the socket descriptor matches
with the socket that is listening to the multicast address. When this is all
true, find the corresponding UDP node by considering solely the IP address and
ignoring the source UDP port.

The main motivation for this comes from an use case where reception is handled
via loopback interface and where the source port is altered when relaying via
the loopback interface. This patch ensures that, regardless of the source port,
as long as messages are sent to the multicast address of interest and from a
predefined IP address, they will be processed.

Since now the structure that holds multicast information is searched when
receiving an UDP datagram, this opened opportunity for the other changes listed
below:
- Move multicast bitrate stats on to multicast info structures.

- Fix aggregation of stats per source address. Now, we can separate the stats of
  each `udpmulticast` instance, even if they are based on the same source
  address (in which case they differ on network interface).

- Remove map of multicast group names and read the name directly from the
  multicast info structure.