Fix processing of partially received tip block
The problem would affect a block partially received within a FEC object that is
flagged as corresponding to a block at the tip of the chain. For example, if a
node is receiving a new block and experiences a temporary outage, in which case
it needs to complete the FEC decoding later based on the repetition data. In the
previous code, it was possible that, when the remainder of this block was
received as a repeated block (not tip block anymore), a previously received
header would be ignored. This patch fixes the problem by always considering a
previously-processed header.