ring-buffer: Add proxy structure for safer reading
The ring buffer's reading flow requires calling GetNextRead() first, then
ConfirmRead() or AbortRead(). In this process, a mutex remains locked.

To ensure that the mutex is unlocked, e.g., on exceptions, use a read proxy
structure. The structure calls GetNextRead() on its constructor and calls
AbortRead() on its destructor in case the read has not been confirmed with
ConfirmRead(). Effectively, the proxy ensures the releasing of the ring-buffer's
mutex.

Add a unit test case covering the proxy and use it on udpnet.