Remove byte/packet rate tracking from ring buffer
The computations are not very accurate. The rates are only updated when the
buffer is read. Hence, if the buffer is idle for a long time, the rate that one
reads is the last rate when the buffer was active. This is not very informative.

Given that the ultimate goal with the buffer statistics is to return them
through the gettxqueueinfo RPC, it is simpler to let the caller of the RPC
figure out the rates based on the byte/packet counts. The resulting
rates (computed by the caller) are more accurate, as they do not depend on the
buffer being active.

This patch drops the rate computations and the corresponding unit test
cases. Given that only byte/packet counts remain, this patch also drops the
enabling of buffer statistics. The counts are cheap enough to leave enabled by
default. Also, they do not depend on parameters anymore (the previous stats
enabling function would take EWMA parameters).